Change http and https launch settings to launch Versions api + Move VersionsContext.cs to models folder
